iio: axi_jesd204b_v51: Remove no-op clock callbacks
The current implementation of the clock registered by the axi_jesd204b_v51
driver does not do anything. The clock is only used for probe ordering.

The clock core provides default implementations for simple clock callbacks
that don't have to do anything.

So remove the no-op clock callbacks, this allows to simplify the code quite
a bit.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>