iio: ad9467: Properly handle clock request errors
The ad9467 driver ignores all errors that might happen when requesting
optional clocks. But there are other errors that can occur other than that
no clock was specified. Most importantly probe deferral, which means a
clock was specified, but it hasn't been registered yet.

Only ignore the error if the error is that no clock has been specified.

For non-optional clocks return the actual error code rather than replacing
it with EPROBE_DEFER. If there is an error other than EPROBE_DEFER is
returned by the clk_get() the clock will never appear regardless how often
it is retried, so there is not point in trying to do so.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>